+++ packages/create-vue-ssg/templates/docs/env.d.ts
@@ -10,3 +10,11 @@ declare module '@codemonster-ru/vueforge-codeblock' {
   import type { DefineComponent } from 'vue'
   export const VfCodeBlock: DefineComponent<Record<string, unknown>, Record<string, unknown>, unknown>
 }
+
+declare module '@/generated/playgroundRegistry.generated' {
+  import type { Component } from 'vue'
+  export const docsVirtualPlaygroundRegistry: Record<string, Component>
+  export const docsVirtualPlaygroundSourceRegistry: Record<string, string>
+  export const docsVirtualPlaygroundSourceLanguageRegistry: Record<string, string>
+  export const docsVirtualPlaygroundSourceFileRegistry: Record<string, string>
+}